requirements.txt
Flask==3.1.3
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.1
python_bcrypt==0.3.2
SQLAlchemy==2.0.46
Werkzeug==3.1.6
